Do not raise bug on invalid interrupt
Since we disable APIC virtualization on entering SEAM mode, we should
remember and reconstruct VMCS GUEST_INTR_STATUS field on exiting SEAM
mode. Also, it is invalid if ISR is on (i.e., interrupt is being
serviced) when SEAMCALL, because every interrupt raised must redirect
the kernel to the handler and be resolved there before executing
SEAMCALL instruction.

Though it doesn't happen on normal execution, when we attach debugger to
TDX module, interrupts are often in invalid state (due to unknown
reasons). For such case, we just ignore interrupt by sending EOI. This
may hang the IO later.

Signed-off-by: Jaewon Hur <hurjaewon@snu.ac.kr>